# GFBackend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the Gator Forum Backend Server",
        "title": "Gator Forum Backend API",
        "termsOfService": "https://github.com/fongziyjun16/SE/tree/backend",
        "contact": {},
        "version": "1.0"
    },
    "host": "http://167.71.166.120:10010",
    "basePath": "/gf/api",
    "paths": {
...
        "/file/delete": {
            "post": {
                "security": [
                    {
                        "ApiAuthToken": []
                    }
                ],
                "description": "need token in cookie, need filename in json",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Static Resource"
                ],
                "summary": "Delete User File, only have permission to delete self data",
                "parameters": [
                    {
                        "description": "filename in post request body",
                        "name": "filename",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.UserFilename"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "\u003cb\u003eSuccess\u003c/b\u003e. Delete Successfully",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseMsg"
                        }
                    },
                    "400": {
                        "description": "\u003cb\u003eFailure\u003c/b\u003e. Bad Parameters or Other",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseMsg"
                        }
                    },
                    "500": {
                        "description": "\u003cb\u003eFailure\u003c/b\u003e. Server Internal Error.",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseMsg"
                        }
                    }
                }
            }
        },
        "/file/download": {
            "post": {
                "security": [
                    {
                        "ApiAuthToken": []
                    }
                ],
                "description": "need token in cookie, need filename in json",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Static Resource"
                ],
                "summary": "User Downloads File, only self data for now",
                "parameters": [
                    {
                        "description": "filename in post request body",
                        "name": "filename",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "\u003cb\u003eSuccess\u003c/b\u003e. Upload Successfully",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseMsg"
                        }
                    },
                    "400": {
                        "description": "\u003cb\u003eFailure\u003c/b\u003e. Bad Parameters or No Enough Space",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseMsg"
                        }
                    },
                    "500": {
                        "description": "\u003cb\u003eFailure\u003c/b\u003e. Server Internal Error.",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseMsg"
                        }
                    }
                }
            }
        },
        "/file/scan": {
            "post": {
                "security": [
                    {
                        "ApiAuthToken": []
                    }
                ],
                "description": "need token in cookie, only get self files",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Static Resource"
                ],
                "summary": "Scan User files",
                "responses": {
                    "201": {
                        "description": "\u003cb\u003eSuccess\u003c/b\u003e. Scan Successfully",
                        "schema": {
                            "$ref": "#/definitions/entity.UserFiles"
                        }
                    },
                    "400": {
                        "description": "\u003cb\u003eFailure\u003c/b\u003e. Bad Parameters.",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseMsg"
                        }
                    },
                    "500": {
                        "description": "\u003cb\u003eFailure\u003c/b\u003e. Server Internal Error.",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseMsg"
                        }
                    }
                }
            }
        },
        "/file/space/info": {
            "post": {
                "security": [
                    {
                        "ApiAuthToken": []
                    }
                ],
                "description": "need token in cookie",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Static Resource"
                ],
                "summary": "Browse User Space Info",
                "responses": {
                    "201": {
                        "description": "\u003cb\u003eSuccess\u003c/b\u003e. Get User Space Info Successfully",
                        "schema": {
                            "$ref": "#/definitions/entity.Space"
                        }
                    },
                    "400": {
                        "description": "\u003cb\u003eFailure\u003c/b\u003e. Bad Parameters or User not exists.",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseMsg"
                        }
                    },
                    "500": {
                        "description": "\u003cb\u003eFailure\u003c/b\u003e. Server Internal Error.",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseMsg"
                        }
                    }
                }
            }
        },
        "/file/space/update": {
            "post": {
                "security": [
                    {
                        "ApiAuthToken": []
                    }
                ],
                "description": "need token in cookie, need target user and  new capacity in json",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Static Resource"
                ],
                "summary": "Expand User Capacity, only admin user can do this",
                "parameters": [
                    {
                        "description": "Username \u0026 New File Total Capacity",
                        "name": "UserNewCapacity",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.UserNewCapacity"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "\u003cb\u003eSuccess\u003c/b\u003e. Update Successfully",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseMsg"
                        }
                    },
                    "400": {
                        "description": "\u003cb\u003eFailure\u003c/b\u003e. Bad Parameters or Other",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseMsg"
                        }
                    },
                    "500": {
                        "description": "\u003cb\u003eFailure\u003c/b\u003e. Server Internal Error.",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseMsg"
                        }
                    }
                }
            }
        },
        "/file/upload": {
            "post": {
                "security": [
                    {
                        "ApiAuthToken": []
                    }
                ],
                "description": "need token in cookie, html file type input element include name attribute with value \"uploadFilename\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Static Resource"
                ],
                "summary": "User Uploads files including images, video etc.",
                "responses": {
                    "201": {
                        "description": "\u003cb\u003eSuccess\u003c/b\u003e. Upload Successfully",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseMsg"
                        }
                    },
                    "400": {
                        "description": "\u003cb\u003eFailure\u003c/b\u003e. Bad Parameters or No Enough Space",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseMsg"
                        }
                    },
                    "500": {
                        "description": "\u003cb\u003eFailure\u003c/b\u003e. Server Internal Error.",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseMsg"
                        }
                    }
                }
            }
        },
        "/file/upload/groupavatar/:groupid": {
            "post": {
                "security": [
                    {
                        "ApiAuthToken": []
                    }
                ],
                "description": "need token in cookie, html file type input element include name attribute with value \"uploadFilename\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Static Resource"
                ],
                "summary": "User Uploads avatar about community that he or she creates",
                "responses": {
                    "201": {
                        "description": "\u003cb\u003eSuccess\u003c/b\u003e. Upload Successfully",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseMsg"
                        }
                    },
                    "400": {
                        "description": "\u003cb\u003eFailure\u003c/b\u003e. Bad Parameters or No Enough Space",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseMsg"
                        }
                    },
                    "500": {
                        "description": "\u003cb\u003eFailure\u003c/b\u003e. Server Internal Error.",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseMsg"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entity.ResponseMsg": {
            "type": "object",
            "properties": {
                "Nickname": {
                    "type": "string",
                    "example": "James Bond"
                },
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "message": {
                    "type": "string",
                    "example": "process successfully"
                },
                "new_community_id": {
                    "type": "integer"
                }
            }
        },
...
        "entity.Space": {
            "type": "object",
            "properties": {
                "capacity": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "used": {
                    "type": "number"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "entity.UserFilename": {
            "type": "object",
            "properties": {
                "Filename": {
                    "type": "string",
                    "example": "gator.jpg"
                }
            }
        },
        "entity.UserFiles": {
            "type": "object",
            "properties": {
                "Filenames": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "\"xxx.jpg\"",
                        "\"xxx.png\"",
                        "\"xxx.gif\""
                    ]
                },
                "Nickname": {
                    "type": "string",
                    "example": "James Bond"
                },
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "message": {
                    "type": "string",
                    "example": "process successfully"
                },
                "new_community_id": {
                    "type": "integer"
                }
            }
        },
...
        "entity.UserNewCapacity": {
            "type": "object",
            "properties": {
                "Capacity": {
                    "type": "number"
                },
                "Username": {
                    "type": "string",
                    "example": "boss"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiAuthToken": {
            "type": "apiKey",
            "name": "token",
            "in": "cookies"
        }
    }
}